Some cmake cleanup
