+++ src/injector/CMakeLists.txt
@@ -1,2 +1,4 @@
-add_executable(injector cxx/main.cpp cxx/injector.cpp cxx/injector.h)
+add_executable(injector main.cpp injector.cpp injector.h)
+target_link_libraries(injector
+    shared)
 add_dependencies(injector injected)
