+++ src/shared/CMakeLists.txt
@@ -1,0 +1,2 @@
+add_library(shared INTERFACE logger.h)
+target_include_directories(shared INTERFACE .)
